Move MOIS sales-library backend classes into worker.v1 package paths
