+++ mois-sales-library-worker/src/main/java/com/marketinghub/mois/bibliotecapaginavenda/worker/v1/config/WorkerProperties.java
@@ -1,4 +1,4 @@
-package com.marketinghub.mois.libraryworker.config;
+package com.marketinghub.mois.bibliotecapaginavenda.worker.v1.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
